requirements.txt
django==2.1
django-allauth==0.37.1
django-cors-headers==2.4.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
Pillow==5.2.0
PyJWT==1.6.4
pytz==2018.4
six==1.11.0
dj-database-url
psycopg2
gunicorn
whitenoise
django-heroku
